requirements.txt
pandas==2.2.2
seaborn==0.13.2
seaborn==0.13.2
pillow==10.3.0import os